requirements.txt
aiohttp
graphene
python-dotenv
